+++ Sources/CredentialCode/Generators/CppCodeGenerator.swift
@@ -246,6 +246,10 @@ struct CppCodeGenerator: CodeGenerator {
         throw CredentialError.notImplemented("External key generation is not yet implemented for C++")
     }
     
+    func generateWithExternalKeySource(credentials: [String: String], encryptionKey: Data) throws -> String {
+        throw CredentialError.notImplemented("External key source generation for C++")
+    }
+    
     func generateKeySource(encryptionKey: Data) throws -> String {
         throw CredentialError.notImplemented("Key source generation is not yet implemented for C++")
     }